refactor: reorganize providers into consistent subfolder structure
Each provider now has a clear capability map visible from directory structure:
- openai/ → chat/ embed/ response/ images/ audio/ codex/
- google/ (was gemini/) → chat/ gemini_code/
- groq/ (extracted from openai/) → chat/ audio_transcription/
- deepseek/ → chat/ (was flat)
- nvidia_nim/ → chat/ embed/ (was flat)
- exa/ → search/ contents/ find_similar/
- firecrawl/ → scrape/ map/ crawl/
- openrouter/ → chat/ (was flat)
- pi/ → cli/ (was flat)
- tavily/ → search/ (was flat)

Structural changes:
- Groq extracted from openai/_config.py + openai/chat/endpoint.py into own provider
- GeminiOAI merged into google/_config.py as GeminiConfigs.CHAT
- CodexConfigs merged into OpenAIConfigs as CODEX_CLI
- All LazyType paths updated to new module locations
- Registry _import_all_providers() updated for all new paths
- token_budget.py provider module paths updated
- Empty placeholder directories cleaned up
- 30+ test files updated for new import paths

5788 tests pass, 0 failures.
