+++ lionagi/providers/deepseek/_config.py
@@ -13,7 +13,7 @@ class DeepSeekConfigs(ProviderConfig, Enum):
         "chat/completions",
         ["chat"],
         EndpointType.API,
-        LazyType("lionagi.providers.deepseek.models:DeepseekChatCompletionsRequest"),
+        LazyType("lionagi.providers.deepseek.chat.models:DeepseekChatCompletionsRequest"),
         "https://api.deepseek.com/v1",
         "bearer",
     )
