--- lionagi/providers/firecrawl/endpoint.py
@@ -1,39 +1,0 @@
-# Copyright (c) 2023-2025, HaiyangLi <quantocean.li at gmail dot com>
-# SPDX-License-Identifier: Apache-2.0
-
-from __future__ import annotations
-
-from lionagi.service.connections.endpoint import Endpoint
-from lionagi.service.connections.endpoint_config import EndpointConfig
-
-from ._config import FirecrawlConfigs
-
-__all__ = ("FirecrawlScrapeEndpoint", "FirecrawlMapEndpoint")
-
-
-@FirecrawlConfigs.SCRAPE.register
-class FirecrawlScrapeEndpoint(Endpoint):
-    def __init__(self, config: EndpointConfig = None, **kwargs):
-        if config is None:
-            from lionagi.config import settings
-
-            kwargs.setdefault(
-                "api_key", settings.FIRECRAWL_API_KEY or "dummy-key-for-testing"
-            )
-            kwargs.setdefault("timeout", 120)
-            kwargs.setdefault("max_retries", 3)
-        super().__init__(config=config, **kwargs)
-
-
-@FirecrawlConfigs.MAP.register
-class FirecrawlMapEndpoint(Endpoint):
-    def __init__(self, config: EndpointConfig = None, **kwargs):
-        if config is None:
-            from lionagi.config import settings
-
-            kwargs.setdefault(
-                "api_key", settings.FIRECRAWL_API_KEY or "dummy-key-for-testing"
-            )
-            kwargs.setdefault("timeout", 120)
-            kwargs.setdefault("max_retries", 3)
-        super().__init__(config=config, **kwargs)
