--- lionagi/providers/nvidia_nim/models.py
@@ -1,8 +1,0 @@
-# Copyright (c) 2023-2026, HaiyangLi <quantocean.li at gmail dot com>
-# SPDX-License-Identifier: Apache-2.0
-
-"""NVIDIA NIM uses OpenAI-compatible request format."""
-
-from lionagi.providers.openai.chat.models import OpenAIChatCompletionsRequest
-
-__all__ = ("OpenAIChatCompletionsRequest",)
