+++ lionagi/providers/pi/_config.py
@@ -13,7 +13,7 @@ class PiConfigs(ProviderConfig, Enum):
         "query_cli",
         ["cli", "code"],
         EndpointType.AGENTIC,
-        LazyType("lionagi.providers.pi.models:PiCodeRequest"),
+        LazyType("lionagi.providers.pi.cli.models:PiCodeRequest"),
     )
