+++ lionagi/service/connections/registry.py
@@ -193,6 +193,8 @@ def _import_all_providers():
         "lionagi.providers.openai.codex.endpoint",
         "lionagi.providers.openai.audio.endpoint",
         "lionagi.providers.openai.images.endpoint",
+        "lionagi.providers.openai.embed.endpoint",
+        "lionagi.providers.openai.response.endpoint",
         # Anthropic
         "lionagi.providers.anthropic.messages.endpoint",
         "lionagi.providers.anthropic.claude_code.endpoint",
@@ -201,19 +203,24 @@ def _import_all_providers():
         "lionagi.providers.ollama.embed.endpoint",
         "lionagi.providers.ollama.generate.endpoint",
         # Search & scraping
-        "lionagi.providers.tavily.endpoint",
-        "lionagi.providers.exa.endpoint",
+        "lionagi.providers.tavily.search.endpoint",
+        "lionagi.providers.exa.search.endpoint",
         "lionagi.providers.exa.contents.endpoint",
         "lionagi.providers.exa.find_similar.endpoint",
-        "lionagi.providers.firecrawl.endpoint",
+        "lionagi.providers.firecrawl.scrape.endpoint",
+        "lionagi.providers.firecrawl.map.endpoint",
         "lionagi.providers.firecrawl.crawl.endpoint",
         # Chat / LLM providers
         "lionagi.providers.perplexity.chat.endpoint",
-        "lionagi.providers.nvidia_nim.endpoint",
-        "lionagi.providers.deepseek.endpoint",
-        "lionagi.providers.gemini.gemini_code.endpoint",
-        "lionagi.providers.pi.endpoint",
-        "lionagi.providers.openrouter.endpoint",
+        "lionagi.providers.nvidia_nim.chat.endpoint",
+        "lionagi.providers.nvidia_nim.embed.endpoint",
+        "lionagi.providers.deepseek.chat.endpoint",
+        "lionagi.providers.google.chat.endpoint",
+        "lionagi.providers.google.gemini_code.endpoint",
+        "lionagi.providers.groq.chat.endpoint",
+        "lionagi.providers.groq.audio_transcription.endpoint",
+        "lionagi.providers.pi.cli.endpoint",
+        "lionagi.providers.openrouter.chat.endpoint",
         # Agentic
         "lionagi.providers.ag2.groupchat.endpoint",
     ]
