+++ tests/service/connections/providers/test_nvidia_nim.py
@@ -8,7 +8,8 @@ import os
 import pytest
 from dotenv import load_dotenv
 
-from lionagi.providers.nvidia_nim.endpoint import NvidiaNimChatEndpoint, NvidiaNimEmbedEndpoint
+from lionagi.providers.nvidia_nim.chat.endpoint import NvidiaNimChatEndpoint
+from lionagi.providers.nvidia_nim.embed.endpoint import NvidiaNimEmbedEndpoint
 from lionagi.service.connections.match_endpoint import match_endpoint
 
 # Create config objects using the new endpoint classes (equivalent to old module-level constants)
